Move serverup stuff to service_checks
